remove all redundant tests and code
